backend/watchpower-api-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.8.2
python-multipart==0.0.6
pytz==2024.1
motor==3.7.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0








